requirements.txt
streamlit
pandas
openpyxl
msal
requests
python-dotenv
